final touchups for control system panel
